backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.104.1
h11==0.16.0
idna==3.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
